Updated samples for PDF Services SDK 3.0.0 release
