+++ src/exportpdf/export-docx-to-pdf-with-ocr-options.js
@@ -12,16 +12,14 @@
 const PDFServicesSdk = require('@adobe/pdfservices-node-sdk');
 
 /**
- * This sample illustrates how to export a PDF file to JPEG.
- * <p>
- * Note that exporting a PDF file to an image format results in a ZIP archive containing one image per page.
+ * This sample illustrates how to export a PDF file to a Word (DOCX) file. The OCR processing is also performed on the input PDF file to extract text from images in the document.
  * <p>
  * Refer to README.md for instructions on how to run the samples.
  */
 
 try {
     // Initial setup, create credentials instance.
-    const credentials =  PDFServicesSdk.Credentials
+    const credentials = PDFServicesSdk.Credentials
         .serviceAccountCredentialsBuilder()
         .fromFile("pdfservices-api-credentials.json")
         .build();
@@ -29,17 +27,21 @@ try {
     //Create an ExecutionContext using credentials and create a new operation instance.
     const executionContext = PDFServicesSdk.ExecutionContext.create(credentials),
         exportPDF = PDFServicesSdk.ExportPDF,
-        exportPdfOperation = exportPDF.Operation.createNew(exportPDF.SupportedTargetFormats.JPEG);
+        exportPdfOperation = exportPDF.Operation.createNew(exportPDF.SupportedTargetFormats.DOCX);
 
     // Set operation input from a source file
-    const input = PDFServicesSdk.FileRef.createFromLocalFile('resources/exportPDFToImageInput.pdf');
+    const input = PDFServicesSdk.FileRef.createFromLocalFile('resources/exportPDFInput.pdf');
     exportPdfOperation.setInput(input);
+
+    // Provide any custom configuration options for the operation.
+    const options = new exportPDF.options.ExportPDFOptions(exportPDF.options.ExportPDFOptions.OCRSupportedLocale.EN_US);
+    exportPdfOperation.setOptions(options);
 
     // Execute the operation and Save the result to the specified location.
     exportPdfOperation.execute(executionContext)
-        .then(result => result.saveAsFile('output/exportPDFToJPEG.zip'))
+        .then(result => result.saveAsFile('output/exportPdfWithOCROptionsOutput.docx'))
         .catch(err => {
-            if(err instanceof PDFServicesSdk.Error.ServiceApiError
+            if (err instanceof PDFServicesSdk.Error.ServiceApiError
                 || err instanceof PDFServicesSdk.Error.ServiceUsageError) {
                 console.log('Exception encountered while executing operation', err);
             } else {
