+++ src/exportpdftoimages/export-pdf-to-jpeg-zip.js
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2021 Adobe
+ * All Rights Reserved.
+ *
+ * NOTICE: Adobe permits you to use, modify, and distribute this file in
+ * accordance with the terms of the Adobe license agreement accompanying
+ * it. If you have received this file from a source other than Adobe,
+ * then your use, modification, or distribution of it requires the prior
+ * written permission of Adobe.
+ */
+
+const PDFServicesSdk = require('@adobe/pdfservices-node-sdk');
+
+/**
+ * This sample illustrates how to export a PDF file to JPEG.
+ * <p>
+ * The resulting file is a ZIP archive containing one image per page of the source PDF file
+ * <p>
+ * Refer to README.md for instructions on how to run the samples.
+ */
+
+try {
+    // Initial setup, create credentials instance.
+    const credentials =  PDFServicesSdk.Credentials
+        .serviceAccountCredentialsBuilder()
+        .fromFile("pdfservices-api-credentials.json")
+        .build();
+
+    //Create an ExecutionContext using credentials and create a new operation instance.
+    const executionContext = PDFServicesSdk.ExecutionContext.create(credentials),
+        exportPDFToImages = PDFServicesSdk.ExportPDFToImages,
+        exportPdfToImagesOperation = exportPDFToImages.Operation.createNew(exportPDFToImages.SupportedTargetFormats.JPEG);
+
+    // Set the output type to create zip of images.
+    exportPdfToImagesOperation.setOutputType(exportPDFToImages.OutputType.ZIP_OF_PAGE_IMAGES);
+
+    // Set operation input from a source file
+    const input = PDFServicesSdk.FileRef.createFromLocalFile('resources/exportPDFToImageInput.pdf');
+    exportPdfToImagesOperation.setInput(input);
+
+    // Execute the operation and Save the result to the specified location.
+    exportPdfToImagesOperation.execute(executionContext)
+        .then(result => result[0].saveAsFile('output/exportPDFToJPEG.zip'))
+        .catch(err => {
+            if(err instanceof PDFServicesSdk.Error.ServiceApiError
+                || err instanceof PDFServicesSdk.Error.ServiceUsageError) {
+                console.log('Exception encountered while executing operation', err);
+            } else {
+                console.log('Exception encountered while executing operation', err);
+            }
+        });
+} catch (err) {
+    console.log('Exception encountered while executing operation', err);
+}
+
